refactor: separate blank encoding
